== ItemID Range & more == *Please don't forget to add the upgrade_15-3athena_svn513 sql-files!
*Corrected guild relay script
-Corrected the value of the variable guildtime to match the value in official script.

*Fixed a few possible crashes.

*Converted some more item related packets to structs.

*Added support for itemids up to ~2 billion.
-Added Concentrated_* potions to test the new range.

+++ sql-files/item_db2.sql
@@ -4,7 +4,7 @@
 
 DROP TABLE IF EXISTS `item_db2`;
 CREATE TABLE `item_db2` (
-  `id` smallint(5) unsigned NOT NULL default '0',
+  `id` int(10) unsigned NOT NULL default '0',
   `name_english` varchar(50) NOT NULL default '',
   `name_japanese` varchar(50) NOT NULL default '',
   `type` tinyint(2) unsigned NOT NULL default '0',
